Add files from the Amazing Race Map App
The application uses the Amazing Race longitude and latitude csv along
with the Amazing Race contest csv file.
The server and ui scripts are provided to run the application. This
application maps locations that teams on The Amazing Race went to over
the first twenty five seasons of the show.